difficutly now implemented. Each time you win or lose, it is adjusted. Added text display for this feature.

+++ src/GridDim.h
@@ -2,7 +2,7 @@
 //TODO: make this better, surely theres a better way to have these consts.
 namespace GridDim
 {
-    inline constexpr int colWidth{ 20 };
+    inline constexpr float colWidth{ 20.f };
     inline constexpr int spriteSize{ 256 };
     inline constexpr int gridSize{ 3 };
     inline constexpr int gridSquares{ gridSize * gridSize };
